requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-binance==1.0.36
groq==0.4.1
cryptography==41.0.7
argon2-cffi==23.1.0
bcrypt==4.1.1
requests==2.31.0
aiohttp==3.9.1
loguru==0.7.2
python-dateutil==2.8.2
Jinja2==3.1.2
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
